refactor: move upgrades into upgrades directory
