+++ packages/core/src/storage/migration/upgrades/0.1-0.2/index.ts
@@ -1,4 +1,4 @@
-import type { Agent } from '../../../agent/Agent'
+import type { Agent } from '../../../../agent/Agent'
 
 import { migrateCredentialRecordToV0_2 } from './credential'
